Move protocols into separate module
