+++ lib/galaxy/datatypes/_protocols.py
@@ -1,0 +1,12 @@
+"""
+Location of protocols used in datatypes
+"""
+
+from typing import Any
+
+from typing_extensions import Protocol
+
+
+class GeneratePrimaryFileDataset(Protocol):
+    extra_files_path: str
+    metadata: Any
