+++ lib/galaxy/datatypes/hdf5.py
@@ -7,10 +7,8 @@ from typing import (
     TYPE_CHECKING,
 )
 
-from galaxy.datatypes.data import (
-    Data,
-    GeneratePrimaryFileDataset,
-)
+from galaxy.datatypes._protocols import GeneratePrimaryFileDataset
+from galaxy.datatypes.data import Data
 from galaxy.datatypes.metadata import MetadataElement
 
 if TYPE_CHECKING:
