+++ lib/galaxy/datatypes/proteomics.py
@@ -11,11 +11,9 @@ from typing import (
 )
 
 from galaxy.datatypes import data
+from galaxy.datatypes._protocols import GeneratePrimaryFileDataset
 from galaxy.datatypes.binary import Binary
-from galaxy.datatypes.data import (
-    GeneratePrimaryFileDataset,
-    Text,
-)
+from galaxy.datatypes.data import Text
 from galaxy.datatypes.sequence import Sequence
 from galaxy.datatypes.sniff import (
     build_sniff_from_prefix,
